+++ relay-dynamic-config/src/project.rs
@@ -42,6 +42,10 @@ pub struct Retentions {
     /// This will determine when they are removed from storage.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub span: Option<RetentionSettings>,
+    /// Retention settings for metrics.
+    /// This will determine when they are removed from storage.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub trace_metric: Option<RetentionSettings>,
 }
 
 impl Retentions {
@@ -62,6 +66,7 @@ impl Retentions {
                     None
                 }
             }
+            TraceItemType::Metric => self.trace_metric.as_ref(),
             _ => None,
         }
     }